+++ CHANGELOG.md
@@ -9,6 +9,9 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 - Feature: Update `Title` component and add new doc.
 - Feature: Add `RocketIcon` and `RocketCircleIcon` components.
 - Feature: Create `ProgressRing` component.
+- Feature: Add `TitleWithStroke` component.
+- Feature: Add `titleModifierStyles` common style for titles.
+- Feature: Add `strokeModifierStyles` common style for horizontal strokes.
 
 ## [2.64.0] - 2020-04-07